+++ client/src/config.js
@@ -1,27 +1,29 @@
-// const CLIENT_ID = process.env.CLIENT_ID || '0oa16k2i3m8hQDTpK5d7';
-// const ISSUER = process.env.ISSUER || 'https://dev-11624122.okta.com/oauth2/default';
-// require("dotenv").config();
-const ISSUER= process.env.ISSUER;
-const CLIENT_ID = process.env.CLIENT_ID;
-//  const ISSUER = `https://${process.env.ISSUER}/oauth2/default`;
-const OKTA_TESTING_DISABLEHTTPSCHECK = process.env.OKTA_TESTING_DISABLEHTTPSCHECK || false;
+
+const ISSUER= process.env.REACT_APP_ISSUER 
+const CLIENT_ID =process.env.REACT_APP_CLIENT_ID 
 const REDIRECT_URI = `${window.location.origin}/login/callback`;
+const OKTA_TESTING_DISABLEHTTPSCHECK = process.env.OKTA_TESTING_DISABLEHTTPSCHECK || false;
+let USE_INTERACTION_CODE = false;
+// if (process.env.REACT_APP_USE_INTERACTION_CODE === 'true') {
+//   USE_INTERACTION_CODE = true;
+// }
 
-
-
-const config = {
+const config= {
     oidc: {
-      clientId: CLIENT_ID,
-      issuer: ISSUER,
-      redirectUri: REDIRECT_URI,
-      scopes: ['openid', 'profile', 'email'],
-      pkce: true,
-      disableHttpsCheck: OKTA_TESTING_DISABLEHTTPSCHECK,
-    },
+        clientId: CLIENT_ID,
+        issuer: ISSUER,
+        redirectUri: REDIRECT_URI,
+        scopes: ['openid', 'profile', 'email'],
+        pkce: true,
+        disableHttpsCheck: OKTA_TESTING_DISABLEHTTPSCHECK,
+        useInteractionCode: USE_INTERACTION_CODE,
+ 
+
+  
     // resourceServer: {
     //   messagesUrl: 'http://localhost:8000/api/messages',
     // },
-
+  }
   };
   
   export default config;
