+++ client/src/index.js
@@ -5,7 +5,8 @@ import ReactDOM from 'react-dom';
 // import { HashRouter as Router } from "react-router-dom";
 import { BrowserRouter as Router } from "react-router-dom";
 // import './index.css';
-import { App } from './App';
+// import { App } from './App';
+import App from './App';
 import registerServiceWorker from './registerServiceWorker';
